Changed viewmodel flow observers binding.
They are called now on first onResume. This is, because is some cases, the view that receives the state could be use a
dialog, that needs to be restored after onRestoreInstance state is called. Otherwise,  the exception "java.lang.IllegalStateException: Can not perform this action after onSaveInstanceState" could be launched.

On restoreInstanceState is called between onStart and onResume, on re-initialization, so binding the views here, guarantees the state of savedInstances has been restored